+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/model/EncryptionAlgorithm.java
@@ -1,0 +1,5 @@
+package ru.dpopkov.knowthenics.model;
+
+public enum EncryptionAlgorithm {
+    BCRYPT, SCRYPT, NOOP
+}